pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robust-ocm"
version = "0.1.0"
description = "Robust Optical Character Recognition and Model Evaluation Toolkit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Robust OCM Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
    "reportlab>=3.6.0",
    "pdf2image>=1.16.0",
    "numpy>=1.21.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
]

[project.scripts]
render = "robust_ocm.render.cli:main"
viz = "robust_ocm.render.visualize:main"
analyze = "robust_ocm.render.analyze:main"
convert-to-omnidoc = "robust_ocm.render.convert_to_omnidoc:main"

[project.urls]
Homepage = "https://github.com/your-org/robust-ocm"
Repository = "https://github.com/your-org/robust-ocm"
Documentation = "https://robust-ocm.readthedocs.io"

[tool.hatch.build.targets.wheel]
packages = ["src/robust_ocm"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88